moved functions from slater_determinants to quadratic_hamiltonian
